requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-telegram-bot